# perf(ts): use Numbers instead of BigInts