@jtbandes 2024-08-13
The fact that v8/chrome is exploring optimizing it is interesting. The issue seems to have had quite a lot of activity; have you learned whether any bigint64 optimizations exist already today and what we need to do to take advantage of them?

> consuming applications such as readers/viewers probably don't care about sub µs resolution

I kinda understand where this is coming from but I'm not sure it applies to Foxglove's use cases (and ROS for that matter). There may be some users who don't care, but as long as there are some that do, I think we need to support original non-rounded values, unless there is some way of dynamically enabling more precision as a user setting?